improvements for usermanagement from cli commands.
- better filtering/search options.
- more robust permissions management.
- refreshed test conventions.
